Fix checkstyle nullpointer error
